chore(SKILL-195): remediation checkpoint on 'feat/SKILL-195-governed-review-evidence-transport' [phase=review loop=review_fix generation=1]
